// frontend/contract.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Id = number;
export type CreatedAt = string;
export type CreatorActorId = string | null;
export type LastEditedAt = string | null;
export type Name = string;
export type JsonSchema = string;
export type DatabaseName = string;
export type TableName = string;

export interface ListFormsResult {
  id: Id;
  created_at: CreatedAt;
  creator_actor_id: CreatorActorId;
  last_edited_at: LastEditedAt;
  name: Name;
  json_schema: JsonSchema;
  database_name: DatabaseName;
  table_name: TableName;
}

/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Name = string;
export type Schema = string;
export type DatabaseName = string;
export type TableName = string;

export interface NewFormParams {
  name: Name;
  schema?: Schema;
  database_name: DatabaseName;
  table_name: TableName;
}

/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type FormName = string;
export type Schema = string;
export type DatabaseName = string;
export type TableName = string;

export interface UiFormParams {
  form_name: FormName;
  schema?: Schema;
  database_name: DatabaseName;
  table_name: TableName;
}

/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Id = number;
export type CreatedAt = string;
export type CreatorActorId = string | null;
export type LastEditedAt = string | null;
export type Name = string;
export type JsonSchema = string;
export type DatabaseName = string;
export type TableName = string;
export type Forms = ListFormsResult[];

export interface UiIndexParams {
  forms: Forms;
}
export interface ListFormsResult {
  id: Id;
  created_at: CreatedAt;
  creator_actor_id: CreatorActorId;
  last_edited_at: LastEditedAt;
  name: Name;
  json_schema: JsonSchema;
  database_name: DatabaseName;
  table_name: TableName;
}

/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Name = string;
export type Schema = string;
export type DatabaseName = string;
export type TableName = string;

export interface ApiNewParams {
  name: Name;
  schema?: Schema;
  database_name: DatabaseName;
  table_name: TableName;
}